chore: renaming analytics index file to wrap
